0.0.70 Pre-alpha - Fixed stylistic errors in the default template; - Fixed display of the comments panel; - Fixed error handling for the interactive form; - Fixed a bug where it was impossible to delete a user through the admin panel; - Added a ban on deleting the superuser (a user that was created during system installation); - Code optimization; - Updated localizations: Russian and English; - Minor bugs fixed;

+++ api/user/delete.handler.php
@@ -25,19 +25,24 @@ if ($system_core->client->is_logged(2)) {
     if (isset($_DELETE['user_id'])) {
       $user_id = (is_numeric($_DELETE['user_id'])) ? (int)$_DELETE['user_id'] : 0;
 
-      if (User::exists_by_id($system_core, $user_id)) {
-        $user = new User($system_core, $user_id);
+      if ($user_id != 1) {
+        if (User::exists_by_id($system_core, $user_id)) {
+          $user = new User($system_core, $user_id);
 
-        $user_is_deleted = $user->delete();
-        if ($user_is_deleted) {
-          $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
-          $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+          $user_is_deleted = $user->delete();
+          if ($user_is_deleted) {
+            $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
+            $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+          } else {
+            $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+            $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+          }
         } else {
-          $handler_message = (!isset($handler_message)) ? (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+          $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USER_ERROR_NOT_FOUND')) : $handler_message;
           $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
         }
       } else {
-        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USER_ERROR_NOT_FOUND')) : $handler_message;
+        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USER_ERROR_CANNOT_REMOVE_SUPERUSER')) : $handler_message;
         $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
       }
     } else {
